+++ TDEngine2/include/ecs/CSystemManager.h
@@ -23,6 +23,9 @@ namespace TDEngine2
 	class IWorld;
 	class ISystem;
 
+
+	TDE2_DECLARE_SCOPED_PTR(ISystem)
+
 	
 	/*!
 		\brief A factory function for creation objects of CSystemManager's type.
@@ -193,6 +196,8 @@ namespace TDEngine2
 			*/
 
 			TDE2_API TSystemId FindSystem(TypeId systemTypeId) override;
+
+			TDE2_API TPtr<ISystem> GetSystem(TSystemId handle) override;
 
 			TDE2_API void ForEachSystem(const std::function<void(TSystemId, const ISystem* const)> action = nullptr) const override;