+++ TDEngine2/include/ecs/ISystemManager.h
@@ -160,6 +160,8 @@ namespace TDEngine2
 
 			TDE2_API virtual TSystemId FindSystem(TypeId systemTypeId) = 0;
 
+			TDE2_API virtual TPtr<ISystem> GetSystem(TSystemId handle) = 0;
+
 			/*!
 				\brief The method allows to iterate over all registered systems including deactivated ones
 			*/